Fix monty-python check.
It turns out that monty-python puts the SA header in the outmost
header and not in the original message header.  So apparently the
mailman moderator message has been classified and not the payload
original mailing list message.  Hmm...  Not optimal but go with it
anyway because of the performance advantages of skipping running
the test here locally.

* bin/mailman-filter: Now check headers instead of payload.