Give processing time before handling spam-new.
When using mutt and saving spam messages to spam-new for queued batch
processing it is possible to make a mistake.  In that case you want to
quickly go to the spam-new folder and remove the message so that it
does not get removed from mailman.  This had been done by using a flag
file to compare against but wasn't working properly.  Convert to using
GNU find specific options --newerXY which perfectly matches what is
needed.

Change the lockfile to be named after the user to ensure that the user
can write the file.  This happens in /tmp so that the associated /tmp
purge will remove any stale locks upon reboot.  This is running on a
friendly system and no attacks against /tmp will happen here.

The spamc apparently returns an (undocumented) 98 when it can't learn
a message because it is already learned.  Handle that non-error case.

Ignore a grep error that is okay to ignore in the event that other
processes are moving files around just as grep looks for them.  This
avoids a spurious error message from grep being seen.

Ensure that the target folder spam-new is a valid folder that includes
cur, tmp, and new.

* bin/mail-learn-spam: Updates as described.